Track Checker added in to Blueprints
Added in a track checker that copies the spline referred to by the train at begin play.

It creates a full spline using all of the points in the connecting splines, making them into one combined spline instead of separate splines. Also disables collision with spline connectors so that it doesn't accidentally jump tracks