Updated Train Movement Methods
The train can now switch between sub-splines inside of the larger complete spline, traversing it as though it were a single spline.

Note for me for later: Consider swapping the application and/or calculation of the movement into FixedUpdate instead of leaving it in Update. Might provide more consistent motion?