Altered map connectors, started Train movement
The map connectors can now take a Name and a reference to the level you want it to load. It will only use one. The Name works in both the editor and the standalone while the level reference only works standalone so far.

I've also started on getting the train moving but have been unable to complete that for this push. Will continue working on it once I'm up again.