Reparented the Engine and the carriage.
Also performed a slew of other changes.
The meshes have been set back to using a single static mesh. Did what I could with the meshes that were there,  found the jittering reduced but not stopped. Suspect the materials may be too complex for something like that to move in that way. Will see what else we can play around with