Cleaned Main Menu, Fixed Train bug in intro
Main Menu had a lot of excess geometry so I've cleared it out. If the artists need it they can go back to previous versions of the level
Also determined the issue with the train in the intro level: there was an old c++ train in there. Removed it because it doesn't have the references it needs from the BP's